pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "study-cli"
version = "0.2.0"
description = "YouTube -> Obsidian Knowledge Compiler"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2024.0",
    "typer>=0.9.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
study = "study.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]
